tiny bit of refactoring common methods into separate file; and a little cleanup
